Extract tool card event reducer
