+++ Tests/QuillCodeParityTests/ParityWorkspaceModelGateTests.swift
@@ -10,6 +10,8 @@ final class ParityWorkspaceModelGateTests: QuillCodeParityTestCase {
         let artifactDocumentPreviewText = try Self.appSourceText(named: "ToolArtifactDocumentPreviewBuilder.swift")
         let artifactTextPreviewText = try Self.appSourceText(named: "ToolArtifactTextPreviewBuilder.swift")
         let transcriptBuilderText = try Self.appSourceText(named: "WorkspaceTranscriptSurfaceBuilder.swift")
+        let toolCardReducerText = try Self.appSourceText(named: "WorkspaceToolCardEventReducer.swift")
+        let toolCardProjectionText = try Self.appSourceText(named: "WorkspaceToolCardProjection.swift")
 
         XCTAssertTrue(toolCardSurfaceText.contains("public struct ToolCardState"), "Tool card surface state should live in a focused surface file.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactState"), "Tool artifact surface state should live in a focused artifact surface file.")
@@ -20,13 +22,20 @@ final class ParityWorkspaceModelGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(artifactImagePreviewText.contains("enum ToolArtifactImagePreviewBuilder"), "Image preview construction should have a focused owner.")
         XCTAssertTrue(artifactDocumentPreviewText.contains("enum ToolArtifactDocumentPreviewBuilder"), "Document preview construction should have a focused owner.")
         XCTAssertTrue(artifactTextPreviewText.contains("enum ToolArtifactTextPreviewBuilder"), "Artifact text-preview file reading should have a focused owner.")
+        XCTAssertTrue(toolCardReducerText.contains("struct WorkspaceToolCardEventReducer"), "Tool-card event state should live in a focused reducer.")
+        XCTAssertTrue(toolCardProjectionText.contains("enum WorkspaceToolCardProjection"), "Tool-card projection copy and formatting should live in a focused projection helper.")
+        XCTAssertTrue(transcriptBuilderText.contains("WorkspaceToolCardEventReducer"), "Transcript projection should delegate tool-card lifecycle state.")
+        XCTAssertTrue(toolCardReducerText.contains("WorkspaceToolCardProjection"), "Tool-card event reduction should delegate card projection details.")
+        XCTAssertTrue(toolCardProjectionText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "Tool-card projection should request artifact text previews through the extracted builder.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactDocumentPreview"), "Document preview contracts should live beside artifact state.")
         XCTAssertTrue(toolArtifactSurfaceText.contains("public struct ToolArtifactImagePreview"), "Image preview contracts should live beside artifact state.")
-        XCTAssertTrue(transcriptBuilderText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "Transcript projection should request artifact text previews through the extracted builder.")
         XCTAssertFalse(modelText.contains("public struct ToolCardState"), "WorkspaceModel should not own tool card surface state.")
         XCTAssertFalse(modelText.contains("public enum ToolCardStatus"), "WorkspaceModel should not own tool card status.")
         XCTAssertFalse(modelText.contains("public struct ToolArtifactState"), "WorkspaceModel should not own tool artifact surface state.")
         XCTAssertFalse(modelText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "WorkspaceModel should not own artifact-preview requests.")
+        XCTAssertFalse(transcriptBuilderText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "Transcript projection should not own artifact-preview requests.")
+        XCTAssertFalse(toolCardReducerText.contains("ToolArtifactTextPreviewBuilder.textPreview"), "Tool-card event reduction should not own artifact-preview requests.")
+        XCTAssertFalse(transcriptBuilderText.contains("private static func approvalReviewCard"), "Transcript projection should not own tool-card approval rendering state.")
         XCTAssertFalse(toolArtifactSurfaceText.contains("private static func documentPreview"), "Tool artifact state should not own document-preview classification.")
         XCTAssertFalse(toolArtifactSurfaceText.contains("private static func isImagePreview"), "Tool artifact state should not own image-preview classification.")
         XCTAssertFalse(toolArtifactSurfaceText.contains("private static func localArtifactFileURL"), "Tool artifact state should not own text-preview file reading.")
